backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
twilio==9.4.1
google-genai==1.5.0
supabase==2.11.0
pydantic==2.10.4
pydantic-settings==2.6.1
httpx==0.28.1
